game_data/scripts/requirements.txt
mysql-connector-sql